[Feat] TargetType 추가 및 NetworkError 분리
